feat: 🎸 增加日志

+++ src/main/java/com/example/todolist/service/UserTaskServiceImpl.java
@@ -2,6 +2,8 @@ package com.example.todolist.service;
 
 import com.example.todolist.dto.Task;
 import com.example.todolist.repository.TaskRepository;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,11 +12,13 @@ import java.util.Optional;
 
 @Service
 public class UserTaskServiceImpl implements UserTaskService{
+    private final Log log = LogFactory.getLog(UserTaskServiceImpl.class);
 
     @Autowired
     TaskRepository taskRepository;
     @Override
     public List<Task> getUserTasks(String userId) {
+        log.info("getUserTasks userId:" + userId);
         return taskRepository.findByUserId(userId);
     }
 
@@ -34,7 +38,11 @@ public class UserTaskServiceImpl implements UserTaskService{
     }
 
     @Override
-    public void deleteUserTask(Long taskId) {
-        taskRepository.deleteById(taskId);
+    public boolean deleteUserTask(Long taskId) {
+        boolean exists = taskRepository.existsById(taskId);
+        if(exists) {
+            taskRepository.deleteById(taskId);
+        }
+        return exists;
     }
 }
